backend/pyproject.toml
[project]
name = "speech-services-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "vosk (>=0.3.45,<0.4.0)",
    "piper-tts (>=1.3.0,<2.0.0)",
    "torch (>=2.8.0,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "torchaudio (>=2.8.0,<3.0.0)",
    "silero-vad (>=6.0.0,<7.0.0)",
    "faster-whisper (>=1.2.0,<2.0.0)",
    "soundfile (>=0.13.1,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
